working plugin tornado plotter
