+++ webviz_subsurface/plugins/_tornado_plotter_fmu/shared_settings/__init__.py
@@ -1,1 +1,4 @@
-from ._filter import MultiFilters, PlotPicker, Selectors, SingleFilters, ViewSettings
+from ._filters import Filters
+from ._plot_picker import PlotPicker
+from ._selectors import Selectors
+from ._view_settings import ViewSettings
