--- webviz_subsurface/plugins/_tornado_plotter_fmu/tornado_widget/__init__.py
@@ -1,1 +1,0 @@
-from ._widget import TornadoWidget
