+++ webviz_subsurface/plugins/_tornado_plotter_fmu/view_elements/_init__.py
@@ -1,4 +1,2 @@
-from ._tornado_data import TornadoData
-from ._tornado_table import TornadoTable
-from ._tornardo_bar_chart import TornadoBarChart
+from ._label import Label
 from ._tornardo_view_element import TornadoViewElement
