+++ webviz_subsurface/plugins/_tornado_plotter_fmu/views/__init__.py
@@ -1,1 +1,1 @@
-from .tornardo_plotter._tornardo_widget import TornadoWidget
+from .tornardo_plotter._tornado_widget import TornadoWidget
